Complete rewrite
Still WIP. Need to work out some things with equations.h and the log.txt errors
I think the AI ate up my big tables. Mode 2 is busted now but I've seen it and it was a smooth 3D surface at 0.03fps. So I worked it up to 30+ fps and now it is ate up.